+++ lib/src/core/builders/dui_icon_builder.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 
-import '../../Utils/basic_shared_utils/lodash.dart';
 import '../../Utils/util_functions.dart';
 import '../../components/dui_icons/icon_helpers/icon_data_serialization.dart';
 import '../../components/dui_widget_scope.dart';
@@ -41,9 +40,12 @@ class DUIIconBuilder extends DUIWidgetBuilder {
 
     iconData ??= getIconData(icondataMap: props['iconData']);
 
-    return Icon(iconData,
-        size: eval<double>(props['iconSize'], context: context),
-        color: eval<String>(props['iconColor'], context: context)
-            .letIfTrue(toColor));
+    return Icon(
+      iconData,
+      size: eval<double>(props['iconSize'], context: context),
+      color: makeColor(
+        eval<String>(props['iconColor'], context: context),
+      ),
+    );
   }
 }