+++ lib/src/core/functions/js_functions.dart
@@ -5,5 +5,21 @@ import 'js_functions_stub.dart'
 abstract class JSFunctions {
   factory JSFunctions() => getJSFunction();
   dynamic callJs(String fnName, dynamic data);
-  fetchJsFile(String path);
+  Future<bool> initFunctions(FunctionInitStrategy strategy);
+  static getFunctionsFileName(int? version) {
+    return version == null ? 'functions.js' : 'functions_v$version.js';
+  }
+}
+
+sealed class FunctionInitStrategy {}
+
+class PreferRemote extends FunctionInitStrategy {
+  final String remotePath;
+  final int? version;
+  PreferRemote(this.remotePath, this.version);
+}
+
+class PreferLocal extends FunctionInitStrategy {
+  final String localPath;
+  PreferLocal(this.localPath);
 }